update encode tests
